Move bar_chart_data_parser, Prune Out Superflourous Method Now in api_connect
